added new bugfixes to data process

+++ migrations/sqls/20201219004151-dataset-input-up.sql
@@ -14,8 +14,8 @@ CREATE TABLE IF NOT EXISTS dataset_db.public.copy_data (
   private boolean null,
   incomplete boolean null,
   last_purchase date null,
-  avg_ticket numeric(5,2) null,
-  last_ticket numeric(5,2) null,
+  avg_ticket float(53) null,
+  last_ticket float(53) null,
   frequent_store varchar(255) null,
   last_store varchar(255) null
 );
